moves chat types to separate folder
